KP-7421 Aestheticize the home page link
It's now always the same colour as generic links, as it's likely not
interesting whether the user has already visited the home page (spoiler
alert: they likely have). It's a bit bigger than ordinary links, but
otherwise it follows the font and style conventions of regular links on
the page.

The size of the navigation link has been chosen so that it does not push
the content downwards but neither can it overlap with the heading, even
if the heading is long and the screen is narrow (as long as "narrow" is
not so narrow that the link text itself requires multiple rows, but
that's highly unlikely as that starts to happen at widths of less than
180 px).